Create Characters_Increasing.cpp
In the file named “CharIncreasing/main.cpp” you need to write a program that reads from standard input and outputs the characters whose ASCII values are strictly larger than the character before it. The one exception to this rule is the newline character and the first character of each line must always be outputted. Characters that are equal or less than their predecessor should be replaced with an underscore character (_) in the output.